server/agentcore-inventory/agents/intake/requirements.txt
# =============================================================================
# IntakeAgent Dependencies
# =============================================================================
# CRITICAL: Keep dependencies MINIMAL for <30s cold start
# =============================================================================

# AWS Bedrock AgentCore Runtime
bedrock-agentcore>=0.1.0

# Google ADK (Agent Development Kit)
google-adk>=0.3.0

# Google Generative AI (for Vision)
google-genai>=0.3.0

# AWS SDK (lazy-loaded)
boto3>=1.35.0

# HTTP client for A2A protocol
httpx>=0.27.0

# X-Ray tracing
aws-xray-sdk>=2.14.0

# Async utilities
anyio>=4.0.0

# NOTE: Avoid heavy dependencies that increase cold start time:
# - NO lxml (use stdlib xml.etree.ElementTree)
# - NO beautifulsoup4
# - NO pandas (use stdlib csv)
# - NO PyPDF2 or pdfplumber (use Vision AI)
